Disabling pull-ups and pull-downs
RPI Pico has pull-downs enabled on GPIOs at start-up.
LPC bus specs say it should be pull-ups on LAD pins (which exist on the motherboard).
To not mess up the LPC bus, all pull-ups and pull-downs are disabled.

LED pin initialization is set in the same manner, making it easy to understand for beginners.